requirements.txt
--find-links wheelhouse
bcrypt==3.1.3
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
Flask-WTF==0.14.2
boto3
nltk
pyodbc
azure-servicebus
git+ssh://git@github.com/The-Pied-Piper/AWS-Manager.git
git+ssh://git@github.com/mwlove/FsLib.git
git+ssh://git@github.com/mwlove/FsORM.git
